+++ playwright/tests/govDocsFilter.spec.ts
@@ -1,13 +1,6 @@
-import { Given, Then, When } from "@cucumber/cucumber";
+import { Then, When } from "@cucumber/cucumber";
 import { expect } from "@playwright/test";
 import { CustomWorld } from "../support/setup";
-
-Given(
-  "I want to filter by government docs",
-  async function (this: CustomWorld) {
-    return await this.page.goto(`${this.parameters.appUrl}`);
-  }
-);
 
 When("I search for swimming", async function (this: CustomWorld) {
   await this.page.locator('[aria-label="Item Search"]').fill("swimming");
@@ -31,6 +24,5 @@ Then(
       .waitFor();
     const textIWant = this.page.locator("a", { hasText: "United States" });
     expect(await textIWant.count()).toBeGreaterThan(3);
-    console.log(textIWant);
   }
 );