+++ playwright/tests/readOnlineTargets.spec.ts
@@ -1,17 +1,6 @@
-import { Given, Then, When } from "@cucumber/cucumber";
+import { Then, When } from "@cucumber/cucumber";
 import { expect } from "@playwright/test";
 import { CustomWorld } from "../support/setup";
-
-Given(
-  "I am on the home page for online targets testing",
-  async function (this: CustomWorld) {
-    return await this.page.goto(`${this.parameters.appUrl}`);
-  }
-);
-
-When("I click Subject: Sub-saharan Africa", async function (this: CustomWorld) {
-  await this.page.locator("text=Subject: Sub-saharan Africa").click();
-});
 
 When(
   "I click Read Online for the first title",
@@ -29,15 +18,7 @@ Then(
     );
   }
 );
-//
-When(
-  "I visit the details page for the title",
-  async function (this: CustomWorld) {
-    return await this.page.goto(
-      `${this.parameters.appUrl}work/01a28167-8c8d-4141-a32f-718539d5c8a4?featured=949699`
-    );
-  }
-);
+
 Then(
   "I click the Read Online button and expect to be on the Hathi Trust website",
   async function (this: CustomWorld) {