+++ src/components/CollectionItemCard/Author.tsx
@@ -1,0 +1,23 @@
+import React from "react";
+import { ApiSearchQuery } from "~/src/types/SearchQuery";
+import Link from "~/src/components/Link/Link";
+
+const Author: React.FC<{ author: string }> = ({ author }) => {
+  if (!author) return null;
+  const query: ApiSearchQuery = {
+    query: `author:${author}`,
+  };
+  return (
+    <Link
+      to={{
+        pathname: "/search",
+        query: query,
+      }}
+      className="link"
+    >
+      {author}
+    </Link>
+  );
+};
+
+export default Author;