+++ src/components/Layout/Layout.test.tsx
@@ -15,9 +15,8 @@ describe("Layout component", () => {
     );
   });
 
-  test("should have main content", () => {
-    const main = screen.getByRole("main");
-    expect(main).toBeInTheDocument();
-    expect(main).toHaveTextContent("Text in layout body");
+  test("should have text in layout body", () => {
+    const text = screen.getByText("Text in layout body");
+    expect(text).toBeInTheDocument();
   });
 });