+++ src/components/Work/Work.tsx
@@ -15,6 +15,11 @@ import {
   Flex,
   Footer,
   TemplateFooter,
+  Text,
+  Card,
+  CardActions,
+  CardContent,
+  CardHeading,
 } from "@nypl/design-system-react-components";
 import {
   joinArrayOfElements,
@@ -29,6 +34,7 @@ import SearchHeader from "../SearchHeader/SearchHeader";
 import { WorkEdition } from "~/src/types/DataModel";
 import Link from "../Link/Link";
 import { MAX_TITLE_LENGTH } from "~/src/constants/editioncard";
+import { PLACEHOLDER_LINK } from "~/src/constants/collection";
 
 const WorkDetail: React.FC<{ workResult: WorkResult; backUrl?: string }> = (
   props
@@ -119,6 +125,42 @@ const WorkDetail: React.FC<{ workResult: WorkResult; backUrl?: string }> = (
                 ></EditionCard>
               </Box>
             </>
+          )}
+          {work.inCollections && work.inCollections.length > 0 && (
+            <Card
+              imageProps={{
+                alt: "Placeholder Cover",
+                aspectRatio: "threeByTwo",
+                isAtEnd: true,
+                size: "large",
+                src: PLACEHOLDER_LINK,
+              }}
+              isCentered
+              backgroundColor="ui.gray.x-light-cool"
+              layout="row"
+              marginTop="l"
+              padding="s"
+            >
+              <CardHeading size="primary" id="row-heading">
+                <Text size="caption" isUppercase marginTop="xs">
+                  <b>Part of Collection</b>
+                </Text>
+                <Box marginTop="m" marginBottom="m">
+                  {work.inCollections[0].title}
+                </Box>
+              </CardHeading>
+              <CardContent marginBottom="l">
+                <Box>{work.inCollections[0].description}</Box>
+              </CardContent>
+              <CardActions width="165px">
+                <Link
+                  linkType="button"
+                  to={"/collection/" + work.inCollections[0].uuid}
+                >
+                  Browse Collection
+                </Link>
+              </CardActions>
+            </Card>
           )}
         </TemplateContentTop>
         <TemplateContentPrimary>